[plugin.video.fosdem] 0.0.5
