+++ plugin.video.fosdem/addon.py
@@ -1,14 +1,12 @@
 # -*- coding: utf-8 -*-
+import sys
 from datetime import datetime, timedelta
-
-import requests
 
 import routing
 from xbmcgui import ListItem
 from xbmcplugin import addDirectoryItem, endOfDirectory, setResolvedUrl, getSetting, setContent
 
-# FIXME: BS4?
-import xml.etree.ElementTree as ET
+from resources.lib.fosdem import fetch_xml, contains_videos
 
 FORMAT_URL = 'https://fosdem.org/{}/schedule/xml'
 FORMATS = ['mp4', 'webm']
@@ -28,17 +26,6 @@ def years():
     # Range does not include the end.
     year += 1
     return range(year - YEARS_SHOWN, year)
-
-
-def fetch_xml(year):
-    r = requests.get(FORMAT_URL.format(year))
-    return ET.fromstring(r.content)
-
-
-def contains_videos(links):
-    videos = list(filter(lambda x: 'video.fosdem.org' in x,
-                  map(lambda x: x.attrib['href'], links)))
-    return videos != []
 
 
 def get_setting_int(name):
@@ -154,4 +141,4 @@ def show_event(year, event_id):
 
 
 if __name__ == '__main__':
-    plugin.run()
+    plugin.run(sys.argv)
