+++ plugin.video.fosdem/addon.xml
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="plugin.video.fosdem" name="FOSDEM Videos" version="0.0.4" provider-name="Jelle van der Waa">
+<addon id="plugin.video.fosdem" name="FOSDEM Videos" version="0.0.5" provider-name="Jelle van der Waa">
     <requires>
         <import addon="xbmc.python" version="2.25.0"/>
         <import addon="script.module.routing" version="0.2.0"/>
-        <import addon="script.module.requests" version="2.3.0"/>
     </requires>
     <extension point="xbmc.python.pluginsource" library="addon.py">
         <provides>video</provides>
@@ -13,6 +12,7 @@
         <description lang="en_GB">This addon provides access to videos published on https://video.fosdem.org</description>
         <language>en</language>
         <platform>all</platform>
+        <reuselanguageinvoker>true</reuselanguageinvoker>
         <license>MIT</license>
         <forum></forum>
         <website>https://fosdem.org</website>
@@ -28,6 +28,8 @@ v0.0.3
 - Added genres to the rooms list menu
 v0.0.4
 - Fixed performance issue retrieving xml data
+v0.0.5
+- Fetch XML data once by caching using reuselangagueinvoker
 	</news>
         <disclaimer lang="en_GB">Unofficial Addon - Use at your own risk</disclaimer>
         <assets>
